requirements.txt
# Bot dependencies
aiogram==3.13.1
aiohttp==3.10.5

# API dependencies
fastapi==0.115.0
uvicorn[standard]==0.31.0
python-multipart==0.0.12
pydantic==2.9.2
pydantic-settings==2.5.2

# Database
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.3
SQLAlchemy==2.0.35

# Utilities
python-dotenv==1.0.1
cryptography==43.0.1

# Export functionality
openpyxl==3.1.5
pandas==2.2.3

# Rate limiting
slowapi==0.1.9
